% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebb_fit_mixture.R
\name{ebb_fit_mixture}
\alias{ebb_fit_mixture}
\alias{ebb_fit_mixture_}
\title{Fit a mixture of beta-binomial distributions}
\usage{
ebb_fit_mixture(tbl, x, n, clusters = 2, iter_max = 10, nstart = 1L,
  vary_size = FALSE, method = "mle", ...)

ebb_fit_mixture_(tbl, x, n, clusters = 2, iter_max = 10, nstart = 1L,
  vary_size = FALSE, method = "mle", ...)
}
\arguments{
\item{tbl}{A table.}

\item{x}{An expression for the number of successes, evaluated within the table.}

\item{n}{An expression for the total number of trials, evaluated within the table.}

\item{clusters}{Number of clusters, default 2}

\item{iter_max}{Maximum number of iterations to perform}

\item{nstart}{Number of random restarts}

\item{vary_size}{Allow each cluster to have a prior probability. Use caution as
this may lead to clusters being lost entirely. If setting this to true, it may
help to use a high number of random restarts.}

\item{method}{Method passed on to \code{\link{ebb_fit_prior}}.}

\item{...}{Extra arguments passed on to \code{\link{ebb_fit_prior}}.}
}
\description{
Fit a mixture of beta-binomial distributions
}
\examples{

library(dplyr)
library(tidyr)
library(purrr)

# simulate some data
set.seed(2017)
sim_data <- data_frame(cluster = 1:2,
                       alpha = c(30, 35),
                       beta = c(70, 15),
                       size = c(300, 700)) \%>\%
  by_row(~ rbeta(.$size, .$alpha, .$beta)) \%>\%
  unnest(p = .out) \%>\%
  mutate(total = round(rlnorm(n(), 5, 2) + 1),
         x = rbinom(n(), total, p))

mm <- ebb_fit_mixture(sim_data, x, total)
mm

# assignments of points to clusters
mm$assignments

# how accurate was it?
mm$assignments \%>\%
  count(cluster, .cluster)

library(ggplot2)
ggplot(mm$assignments, aes(x / total, fill = .cluster)) +
  geom_histogram()

}

