% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebb_mixture-tidiers.R
\name{ebb_mixture_tidiers}
\alias{augment.ebb_mixture}
\alias{ebb_mixture_tidiers}
\alias{glance.ebb_mixture}
\alias{tidy.ebb_mixture}
\title{Tidy a mixture of beta-binomials}
\usage{
\method{tidy}{ebb_mixture}(x, ...)

\method{augment}{ebb_mixture}(x, data, ...)

\method{glance}{ebb_mixture}(x, ...)
}
\arguments{
\item{x}{An ebb_mixture object, such as returned by \code{\link{ebb_fit_mixture}}.}

\item{...}{Extra arguments, not used}

\item{data}{The original data that the mixture model was fit on}
}
\value{
\code{tidy} returns a tbl_df with one row for each cluster,
with columns
  \item{cluster}{A character vector with cluster IDs}
  \item{alpha}{alpha shape parameter for this cluster}
  \item{beta}{beta shape parameter for this cluster}
  \item{mean}{alpha / beta, the mean of the beta distribution}
  \item{size}{Size of this cluster in the final assignments}
  \item{probability}{Prior probability of being in this cluster}

\code{glance} returns a one-row data frame with columns:
  \item{iter}{Number of iterations}
  \item{logLik}{Log-likelihood of the model, given the final
  "hard" assignments to each cluster}
  \item{AIC}{AIC of the model}
  \item{BIC}{BIC of the model}
}
\description{
Tidy a mixture model fit by \code{\link{ebb_fit_mixture}}. \code{tidy} returns
in the hyperparameters fit for each cluster, while \code{glance} returns the
likelihood and other parameters for the full mixture.
}

