% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebb_prior-tidiers.R
\name{ebb_prior_tidiers}
\alias{augment.ebb_prior}
\alias{ebb_prior_tidiers}
\alias{glance.ebb_prior}
\alias{tidy.ebb_prior}
\title{Tidy the results of beta-binomial estimation}
\usage{
\method{tidy}{ebb_prior}(x, ...)

\method{augment}{ebb_prior}(x, data, conf.level = 0.95, newdata = NULL, ...)

\method{glance}{ebb_prior}(x, ...)
}
\arguments{
\item{x}{An object of class "ebb_prior", representing a beta-binomial distribution.}

\item{...}{Extra arguments (not used)}

\item{data}{For augment, the original data that should be appended to.}

\item{conf.level}{For augment, the level of the credible intervals for each
observation. If NULL, do not add credible intervals.}

\item{newdata}{For augment, a new data set.}
}
\value{
\code{augment} returns the original data frame, with additional
columns corresponding to the new posterior parameters of each observation.

\code{glance} works only on ebb_prior objects fit with "mm" or "mle": that is,
those that have a single alpha and beta for the entire dataset. It returns
those parameters:
  \item{alpha}{Estimated alpha (shape1) parameter}
  \item{beta}{Estimated beta (shape2) parameter}
  \item{mean}{alpha / (alpha + beta): the prior mean}
}
\description{
Tidiers for ebb_prior objects, which store hyperparameters for a prior fit on
some binomial data.
}

