% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ebbinom_tidiers.R
\name{ebbinom_tidiers}
\alias{augment.ebbinom}
\alias{ebbinom_tidiers}
\alias{glance.ebbinom}
\title{Tidy the results of empirical Bayes shrinkage}
\usage{
\method{augment}{ebbinom}(x, data, ...)

\method{glance}{ebbinom}(x, ...)
}
\arguments{
\item{x}{An object of class "ebbinom"}

\item{data}{For augment, the original data that
should be appended to}

\item{...}{Extra arguments (not used)}
}
\value{
\code{augment.ebbinom} returns the original
data frame (if provided as the \code{data} argument),
with additional columns corresponding
to each of the ones in \code{x$table}, including the
new alpha and beta posterior parameters for each
observation.

\code{glance.ebbinom} returns the prior as it has been
estimated from the full dataset, including
  \item{alpha}{Estimated alpha (shape1) parameter}
  \item{beta}{Estimated beta (shape2) parameter}
  \item{prior.estimate}{alpha / (alpha + beta): the prior mean}
}
\description{
Tidy the results of empirical Bayes shrinkage
}

