% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate_eb_proportion.R
\name{estimate_eb_proportion}
\alias{estimate_eb_proportion}
\title{Estimate proportions of various success/total counts}
\usage{
estimate_eb_proportion(x, n, prior_filter = TRUE, cred_level = 0.05)
}
\arguments{
\item{x}{Vector of successes}

\item{n}{Vector of totals}

\item{prior_filter}{A logical vector indicating which
values should be used for computing the prior}

\item{cred_level}{Level of credible interval to
compute. If NULL, do not compute intervals.}
}
\value{
An "ebbinom" object, which contains a data.frame
\code{table} with the columns:
  \item{alpha1}{Posterior alpha (shape1) parameter}
  \item{beta1}{Posterior beta (shape2) parameter}
  \item{estimate}{Posterior shrunken estimate}
  \item{raw.estimate}{Estimate without shrinkage (success / total)}
  \item{conf.low}{Lower bound of credible interval}
  \item{conf.high}{Upper bound of credible interval}

It also contains a data frame \code{prior} with the estimated
parameters for the beta distribution.
}
\description{
Estimate proportions of various success/total counts
}
\examples{
library(dplyr)

set.seed(2015)

true_prob <- rbeta(1000, 10, 50)
total <- round(rlnorm(1000, 5, 2)) + 1
success <- rbinom(1000, total, true_prob)

result <- estimate_eb_proportion(success, total, total > 100)

prior <- glance(result)

library(ggplot2)

ggplot(result$table, aes(raw.estimate, estimate, color = log10(total))) +
  geom_point() +
  geom_abline(color = "red") +
  geom_hline(yintercept = prior$prior.estimate,
             color = "red", lty = 2)
}
\seealso{
ebbinom_tidiers
}

