% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_beta.R
\name{geom_beta}
\alias{geom_beta}
\title{Add a layer showing a beta estimation}
\usage{
geom_beta(estimates, color = "red", ...)
}
\arguments{
\item{estimates}{A data.frame returned from an estimation
function, with alpha and beta parameters}

\item{color}{Color of overlaid line}

\item{...}{Extra arguments passed on to stat_function}
}
\description{
Add a layer showing a beta estimation
}

