% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.R
\name{h}
\alias{h}
\title{Calculate the probability one random beta variable is greater than another}
\source{
The exact version comes from Evan Miller and Chris Stucchio:
\url{https://www.chrisstucchio.com/blog/2014/bayesian_ab_decision_rule.html}

John D. Cook lays out the normal approximation here:
\url{http://www.johndcook.com/fast_beta_inequality.pdf}.
}
\usage{
h(a, b, c, d, approx = FALSE, log_h = FALSE)
}
\arguments{
\item{a}{alpha parameter for first Beta}

\item{b}{beta parameter for second Beta}

\item{c}{alpha parameter for first Beta}

\item{d}{beta parameter for second Beta}

\item{approx}{whether to use a normal approximation to the beta}

\item{log_h}{whether to return \code{log(h(a, b, c, d))} rather than
\code{h(a, b, c, d)}}
}
\description{
Find the probability that \code{Beta(a, b) > Beta(c, d)}, using either
an exact solution or a normal approximation. Vectorized across
a, b, c and d.
}
\details{
The "exact" solution is exact only for integer values of c.
}
\examples{

h(60, 40, 50, 50)
h(60, 40, 50, 50, approx = TRUE)

# check against random simulation:
mean(rbeta(1e5, 60, 40) > rbeta(1e5, 50, 50))

# It is vectorized across one or multiple arguments
a <- 1:19
h(a, 20 - a, 10, 10)
plot(a, h(a, 20 - a, 10, 10))

}

