% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebb_mixture-methods.R
\name{logLik.ebb_mixture}
\alias{logLik.ebb_mixture}
\title{Log-likelihood of an ebb_mixture object}
\usage{
\method{logLik}{ebb_mixture}(object, ...)
}
\arguments{
\item{object}{An ebb_mixture object returned by \code{\link{ebb_fit_mixture}}.}

\item{...}{Extra arguments, not used.}
}
\description{
Log-likelihood of an ebb_mixture object
}
