% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebb_prior-methods.R
\name{logLik.ebb_prior}
\alias{logLik.ebb_prior}
\title{Log-likelihood of an ebb_prior object}
\usage{
\method{logLik}{ebb_prior}(object, ...)
}
\arguments{
\item{object}{An "ebb_prior" object}

\item{...}{Extra arguments, not used}
}
\description{
Log-likelihood of an ebb_prior object
}

