% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebb_prior-methods.R
\name{model.frame.ebb_prior}
\alias{model.frame.ebb_prior}
\title{Extract the data frame with the variables needed to evaluate an
empirical Bayes model fit}
\usage{
\method{model.frame}{ebb_prior}(formula, ...)
}
\arguments{
\item{formula}{An ebb_prior object}

\item{...}{Extra arguments, not used}
}
\description{
Extract the data frame with the variables needed to evaluate an
empirical Bayes model fit
}
