% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebb_mixture-methods.R
\name{print.ebb_mixture}
\alias{print.ebb_mixture}
\title{Print an ebb_mixture object}
\usage{
\method{print}{ebb_mixture}(x, ...)
}
\arguments{
\item{x}{An ebb_mixture object returned by \code{\link{ebb_fit_mixture}}.}

\item{...}{Extra arguments, not used.}
}
\description{
Print an ebb_mixture object
}
