% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebb_prior-methods.R
\name{print.ebb_prior}
\alias{print.ebb_prior}
\title{Print an ebb_prior object}
\usage{
\method{print}{ebb_prior}(x, ...)
}
\arguments{
\item{x}{An ebb_prior object}

\item{...}{Extra arguments, not used}
}
\description{
Print an ebb_prior object
}

