% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebb_prior-tidiers.R
\docType{import}
\name{reexports}
\alias{augment}
\alias{glance}
\alias{reexports}
\alias{tidy}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom]{tidy}}, \code{\link[broom]{augment}}, \code{\link[broom]{glance}}}
}}
\keyword{internal}

