% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ebb_estimate.R
\name{add_ebb_estimate}
\alias{add_ebb_estimate}
\alias{add_ebb_estimate_}
\title{Perform empirical Bayes shrinkage on a data frame}
\usage{
add_ebb_estimate(tbl, x, n, cred_level = 0.95, prior_subset = TRUE, ...)

add_ebb_estimate_(tbl, x, n, prior_subset = TRUE, cred_level = 0.05, ...)
}
\arguments{
\item{tbl}{A table.}

\item{x}{Column containing number of successes.}

\item{n}{Column containing totals.}

\item{cred_level}{Level of credible interval to compute. If NULL,
do not compute intervals.}

\item{prior_subset}{An expression evaluating to a logical vector
indicating which values should be used for computing the prior.}

\item{...}{Extra arguments passed on to \code{\link{ebb_fit_prior}},
such as \code{method}.}
}
\value{
The original table, with several columns added based on
empirical Bayes shrinkage:
  \item{.alpha1}{Posterior alpha (shape1) parameter}
  \item{.beta1}{Posterior beta (shape2) parameter}
  \item{.fitted}{Posterior shrunken estimate}
  \item{.raw}{Estimate without shrinkage (success / total)}
  \item{.low}{Lower bound of credible interval}
  \item{.high}{Upper bound of credible interval}
}
\description{
Add columns to a data frame representing empirical Bayes shrinkage
towards an estimated beta prior. This calls the \code{\link{ebb_fit_prior}}
function to fit the prior, then adds the columns using
\code{\link{augment.ebb_prior}}. It is thus a useful wrapper when you're
not interested in the prior itself, but only in performing shrinkage
on data.
}
\examples{

library(dplyr)
set.seed(2017)

# simulate 200 random examples from a beta-binomial
obs <- 200
dat <- data_frame(prob = rbeta(obs, 10, 50),
                  total = round(rlnorm(obs, 4, 2)) + 1,
                  x = rbinom(obs, total, prob))

result <- add_ebb_estimate(dat, x, total)
result

# visualize the shrinkage towards the prior mean
library(ggplot2)
ggplot(result, aes(.raw, .fitted, color = log10(total))) +
  geom_point() +
  geom_abline(color = "red")

}
\seealso{
ebb_prior_tidiers
}

