% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate_beta_binom.R
\name{estimate_beta_binom}
\alias{estimate_beta_binom}
\title{Estimate the shape parameters of a beta-binomial distribution}
\usage{
estimate_beta_binom(successes, totals)
}
\arguments{
\item{successes}{Vector of successes}

\item{totals}{Vector of totals}
}
\description{
Estimate the shape parameters of a beta-binomial distribution
}

