% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/beta_mm.R
\name{estimate_beta_mm}
\alias{estimate_beta_mm}
\title{Estimate parameters for the beta distribution using the
method of moments.}
\usage{
estimate_beta_mm(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A data.frame with two values: alpha and beta.
}
\description{
Estimate the two shape parameters (here called alpha and
beta) for the beta distribution using the method of
moments.
}

