% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_beta.R
\name{geom_beta}
\alias{geom_beta}
\title{Add a layer showing a beta estimation}
\usage{
geom_beta(estimates, color = "red", ...)
}
\arguments{
\item{estimates}{A data.frame returned from an estimation
function, with alpha and beta parameters}

\item{color}{Color of overlaid line}

\item{...}{Extra arguments passed on to stat_function}
}
\description{
Add a layer showing a beta estimation
}
\examples{
set.seed(2015)
probs <- rbeta(100, 10, 50)
total <- round(rlnorm(100, 5, 2)) + 1
successes <- rbinom(100, total, probs)

bb <- estimate_beta_binom(successes, total)
bb

library(ggplot2)

dat <- data.frame(successes, total)
ggplot(dat, aes(successes / total)) +
    geom_histogram(aes(y = ..density..)) +
    geom_beta(bb)
}

